feat: add responsive header with HamburgerSelect
